Weapons and Better Collisions
I've improved the collisions detection system and also improved the wall
collision resolutions. (the wall resolutions aren't perfect, but they're
a ton better)

Also, I've built upon the weapon class. The weapons can either have an
owner or not, wen they do not they appear with an "on ground" texture.
When the player collides with a weapon the weapon's owner is set to the
player and it will respond to clicks for shooting. All weapons track
ther own ammo clips and "gun type" (currently their are SMG and shotgun
types). The Projectile class will travel at a given speed along a given
trajectory and automatically modifies it's trajectory by plus or minus
some random "stray", the amount of which is dependent on the gun type.

I've also implemented DrawOrder for some of the objects. XNA
automatically manages the sorting of the sprites in the SpriteBatch for
us, so all we have to do is set the DrawOrder value and XNA will do the
rest.